Removing project.lock.json from GitHub
project.lock.json and Source Control

There are two schools of thought regarding these lock files and source control. They correspond to two similar camps regarding checking in dependencies themselves (very commonly done before NuGet became popular). In the first case, the goal is to commit to source control the specific dependencies (either the DLLs themselves, or a very precise list of them via project.lock.json) used by the application at that moment in time. In the second case, the goal is to allow the application to specify its minimum dependencies, but allow it to pull in the latest updated versions of these dependencies (within the ranges specified in project.json). Both sides have pros and cons. It’s worth noting that if you use the default visual studio .gitignore file on GitHub today, project.lock.json is specified to not be committed to source control.